+++ services/email/src/oci_cli_email/email_cli_extended.py
@@ -1,0 +1,6 @@
+# coding: utf-8
+# Copyright (c) 2016, 2019, Oracle and/or its affiliates. All rights reserved.
+
+from oci_cli_email.generated import email_cli
+
+email_cli.email_root_group.help = "Email Delivery Service CLI"