+++ services/object_storage/src/oci_cli_object_storage/objectstorage_cli_extended.py
@@ -113,6 +113,7 @@ cli_util.rename_command(objectstorage_cli.os_root_group, objectstorage_cli.names
 cli_util.rename_command(objectstorage_cli.namespace_group, objectstorage_cli.get_namespace_metadata, "get-metadata")
 cli_util.rename_command(objectstorage_cli.namespace_group, objectstorage_cli.update_namespace_metadata, "update-metadata")
 cli_util.rename_command(objectstorage_cli.os_root_group, objectstorage_cli.preauthenticated_request_group, "preauth-request")
+cli_util.rename_command(objectstorage_cli.work_request_log_entry_group, objectstorage_cli.list_work_request_logs, "list")
 
 objectstorage_cli.os_root_group.help = "Object Storage Service CLI"
 objectstorage_cli.os_root_group.short_help = "Object Storage Service"